# apps/doc-engine/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.AddTagsRequest:
    properties:
      tagIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - tagIds
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.CloneTemplateRequest:
    properties:
      newTitle:
        maxLength: 255
        minLength: 1
        type: string
      targetFolderId:
        type: string
      versionId:
        type: string
    required:
    - newTitle
    - versionId
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.CreateTemplateRequest:
    properties:
      contentStructure:
        description: Initial content for the first version
        items:
          type: integer
        type: array
      folderId:
        type: string
      isPublicLibrary:
        type: boolean
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.FolderResponse:
    properties:
      childFolderCount:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      parentId:
        type: string
      templateCount:
        type: integer
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.FormatConfigResponse:
    properties:
      default:
        type: string
      options:
        items:
          type: string
        type: array
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InjectableResponse:
    properties:
      createdAt:
        type: string
      dataType:
        type: string
      description:
        type: string
      formatConfig:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.FormatConfigResponse'
      id:
        type: string
      isGlobal:
        type: boolean
      key:
        type: string
      label:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      sourceType:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListTemplatesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateListItemResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TagResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TagSimpleResponse:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateCreateResponse:
    properties:
      initialVersion:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionResponse'
      template:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateResponse'
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateListItemResponse:
    properties:
      createdAt:
        type: string
      folderId:
        type: string
      hasPublishedVersion:
        type: boolean
      id:
        type: string
      isPublicLibrary:
        type: boolean
      publishedVersionNumber:
        type: integer
      scheduledVersionCount:
        type: integer
      tags:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TagSimpleResponse'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      versionCount:
        type: integer
      workspaceId:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateResponse:
    properties:
      createdAt:
        type: string
      folderId:
        type: string
      id:
        type: string
      isPublicLibrary:
        type: boolean
      title:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionDetailResponse:
    properties:
      archivedAt:
        type: string
      archivedBy:
        type: string
      contentStructure:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      injectables:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionInjectableResponse'
        type: array
      name:
        type: string
      publishedAt:
        type: string
      publishedBy:
        type: string
      scheduledArchiveAt:
        type: string
      scheduledPublishAt:
        type: string
      signerRoles:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionSignerRoleResponse'
        type: array
      status:
        type: string
      templateId:
        type: string
      updatedAt:
        type: string
      versionNumber:
        type: integer
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionInjectableResponse:
    properties:
      createdAt:
        type: string
      defaultValue:
        type: string
      definition:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.InjectableResponse'
      id:
        type: string
      isRequired:
        type: boolean
      templateVersionId:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionResponse:
    properties:
      archivedAt:
        type: string
      archivedBy:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      publishedAt:
        type: string
      publishedBy:
        type: string
      scheduledArchiveAt:
        type: string
      scheduledPublishAt:
        type: string
      status:
        type: string
      templateId:
        type: string
      updatedAt:
        type: string
      versionNumber:
        type: integer
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionSignerRoleResponse:
    properties:
      anchorString:
        type: string
      createdAt:
        type: string
      id:
        type: string
      roleName:
        type: string
      signerOrder:
        type: integer
      templateVersionId:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionSummaryResponse:
    properties:
      archivedAt:
        type: string
      archivedBy:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      injectables:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionInjectableResponse'
        type: array
      name:
        type: string
      publishedAt:
        type: string
      publishedBy:
        type: string
      scheduledArchiveAt:
        type: string
      scheduledPublishAt:
        type: string
      signerRoles:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionSignerRoleResponse'
        type: array
      status:
        type: string
      templateId:
        type: string
      updatedAt:
        type: string
      versionNumber:
        type: integer
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateWithAllVersionsResponse:
    properties:
      createdAt:
        type: string
      folder:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.FolderResponse'
      folderId:
        type: string
      id:
        type: string
      isPublicLibrary:
        type: boolean
      tags:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TagResponse'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      versions:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionSummaryResponse'
        type: array
      workspaceId:
        type: string
    type: object
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateWithDetailsResponse:
    properties:
      createdAt:
        type: string
      folder:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.FolderResponse'
      folderId:
        type: string
      id:
        type: string
      isPublicLibrary:
        type: boolean
      publishedVersion:
        $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateVersionDetailResponse'
      tags:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TagResponse'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.UpdateTemplateRequest:
    properties:
      folderId:
        description: Use "root" to move template to root folder
        type: string
      isPublicLibrary:
        type: boolean
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Document Assembly System API - Template management and document generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Doc Engine API
  version: "1.0"
paths:
...
  /api/v1/content/templates:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Filter by folder ID. Use 'root' to get only root-level templates
          (no folder)
        in: query
        name: folderId
        type: string
      - description: Filter by published status
        in: query
        name: hasPublishedVersion
        type: boolean
      - collectionFormat: csv
        description: Filter by tag IDs
        in: query
        items:
          type: string
        name: tagIds
        type: array
      - description: Search by title
        in: query
        name: search
        type: string
      - description: Limit results
        in: query
        name: limit
        type: integer
      - description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListTemplatesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: List templates
      tags:
      - Templates
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Create template
      tags:
      - Templates
  /api/v1/content/templates/{templateId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Delete template
      tags:
      - Templates
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateWithDetailsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get template
      tags:
      - Templates
    put:
      consumes:
      - application/json
      description: Updates a template's metadata. Use folderId="root" to move the
        template to the root folder.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      - description: Template data (folderId can be a folder UUID or 'root' to move
          to root)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Update template
      tags:
      - Templates
  /api/v1/content/templates/{templateId}/all-versions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateWithAllVersionsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get template with all versions
      tags:
      - Templates
  /api/v1/content/templates/{templateId}/clone:
    post:
      consumes:
      - application/json
      description: Clones a template using the content from a specific version (identified
        by versionId in request body). The versionId must belong to the specified
        templateId.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      - description: Clone data (versionId is required and must belong to the template)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.CloneTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.TemplateCreateResponse'
        "400":
          description: Bad request (invalid versionId, version doesn't belong to template,
            validation error)
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Template or version not found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "409":
          description: Template title already exists
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Clone template from specific version
      tags:
      - Templates
  /api/v1/content/templates/{templateId}/tags:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      - description: Tag IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.AddTagsRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Add tags to template
      tags:
      - Templates
  /api/v1/content/templates/{templateId}/tags/{tagId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Enable sandbox mode (operates on sandbox workspace)
        in: header
        name: X-Sandbox-Mode
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      - description: Tag ID
        in: path
        name: tagId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Remove tag from template
      tags:
      - Templates
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"